AC-856: extract harness codegen, skill lesson, and snapshot mixins from ArtifactStore
Pure move of three method groups into sibling mixin modules, following the
existing PlaybookApprovalMethods/ArtifactGenerationPersistenceMethods/
ArtifactWriteMethods precedent. Shrinks artifacts.py from 1267 to 991 lines.

Shared helpers used by both the harness codegen group and the out-of-scope
tools group (_list_python_modules, _render_python_context,
_wrap_generated_module, _persist_generated_modules) stay on ArtifactStore to
avoid duplication. HarnessCodegenMethods uses the stub-method pattern
(matching sqlite_store's mixins) instead of a Host Protocol, since its
public methods call sibling methods within the same mixin and a
self-referential Protocol hit a mypy limitation for external callers.
