+++ autocontext/src/autocontext/storage/artifact_harness_codegen.py
@@ -1,0 +1,93 @@
+"""Harness validator codegen methods for ArtifactStore."""
+
+from __future__ import annotations
+
+import re
+from collections.abc import Callable
+from pathlib import Path
+from typing import Any
+
+
+class HarnessCodegenMethods:
+    def _scenario_dir(self, scenario_name: str) -> Path:
+        raise NotImplementedError
+
+    def _list_python_modules(self, directory: Path) -> list[str]:
+        raise NotImplementedError
+
+    def _render_python_context(
+        self,
+        directory: Path,
+        *,
+        empty_message: str,
+        name_prefix: str = "",
+    ) -> str:
+        raise NotImplementedError
+
+    def _persist_generated_modules(
+        self,
+        directory: Path,
+        generation_index: int,
+        specs: list[dict[str, Any]],
+        *,
+        kind: str,
+        header_template: str,
+        name_validator: Callable[[str], str] | None = None,
+    ) -> list[str]:
+        raise NotImplementedError
+
+    def harness_dir(self, scenario_name: str) -> Path:
+        """Return the harness directory: knowledge/<scenario>/harness/"""
+        return self._scenario_dir(scenario_name) / "harness"
+
+    @staticmethod
+    def _validate_harness_name(name: str) -> str:
+        """Validate harness module name and prevent path traversal."""
+        candidate = name.strip()
+        if not re.fullmatch(r"[a-zA-Z_][a-zA-Z0-9_]*", candidate):
+            raise ValueError(f"invalid harness name: {name!r}")
+        return candidate
+
+    def persist_harness(
+        self,
+        scenario_name: str,
+        generation_index: int,
+        specs: list[dict[str, Any]],
+    ) -> list[str]:
+        """AST-validate and write harness .py files, archiving old versions."""
+        return self._persist_generated_modules(
+            self.harness_dir(scenario_name),
+            generation_index,
+            specs,
+            kind="harness",
+            header_template="Harness validator generated by architect in generation {generation_index}.",
+            name_validator=self._validate_harness_name,
+        )
+
+    def write_harness(self, scenario_name: str, name: str, source: str) -> Path:
+        """Write a single harness file to knowledge/<scenario>/harness/<name>.py."""
+        safe_name = self._validate_harness_name(name)
+        h_dir = self.harness_dir(scenario_name)
+        h_dir.mkdir(parents=True, exist_ok=True)
+        target = h_dir / f"{safe_name}.py"
+        target.write_text(source, encoding="utf-8")
+        return target
+
+    def read_harness(self, scenario_name: str, name: str) -> str | None:
+        """Read a harness file by name, or None if not found."""
+        safe_name = self._validate_harness_name(name)
+        target = self.harness_dir(scenario_name) / f"{safe_name}.py"
+        if not target.exists():
+            return None
+        return target.read_text(encoding="utf-8")
+
+    def list_harness(self, scenario_name: str) -> list[str]:
+        """List all harness file names for a scenario (sorted, without .py extension)."""
+        return self._list_python_modules(self.harness_dir(scenario_name))
+
+    def read_harness_context(self, scenario_name: str) -> str:
+        """Read harness validator files as markdown context for prompts."""
+        return self._render_python_context(
+            self.harness_dir(scenario_name),
+            empty_message="No harness validators available.",
+        )
