+++ autocontext/src/autocontext/storage/artifacts.py
@@ -4,7 +4,6 @@ import ast
 import json
 import logging
 import os
-import re
 from collections.abc import Callable
 from importlib import import_module
 from pathlib import Path
@@ -27,6 +26,9 @@ from autocontext.knowledge.hint_volume import HintManager, HintVolumePolicy
 from autocontext.knowledge.lessons import LessonStore
 from autocontext.knowledge.mutation_log import MutationEntry, MutationLog
 from autocontext.storage.artifact_generation_persistence import ArtifactGenerationPersistenceMethods
+from autocontext.storage.artifact_harness_codegen import HarnessCodegenMethods
+from autocontext.storage.artifact_skill_lessons import SkillLessonMethods
+from autocontext.storage.artifact_snapshots import SnapshotMethods
 from autocontext.storage.artifact_write_methods import ArtifactWriteMethods
 from autocontext.storage.blob_integration import BlobAwareWriter, mirror_path_append_bytes, mirror_path_bytes
 from autocontext.storage.buffered_writer import BufferedWriter
@@ -51,10 +53,18 @@ class DictSerializable(Protocol):
 
     def to_dict(self) -> dict[str, Any]: ...
 
+
 EMPTY_PLAYBOOK_SENTINEL = "No playbook yet. Start from scenario rules and observation."
 
 
-class ArtifactStore(PlaybookApprovalMethods, ArtifactGenerationPersistenceMethods, ArtifactWriteMethods):
+class ArtifactStore(
+    PlaybookApprovalMethods,
+    ArtifactGenerationPersistenceMethods,
+    ArtifactWriteMethods,
+    HarnessCodegenMethods,
+    SkillLessonMethods,
+    SnapshotMethods,
+):
     def __init__(
         self,
         runs_root: Path,
@@ -237,13 +247,10 @@ class ArtifactStore(PlaybookApprovalMethods, ArtifactGenerationPersistenceMethod
         start = content.find(marker)
         if start == -1:
             return ""
-        after = content[start + len(marker):]
+        after = content[start + len(marker) :]
         next_heading = after.find("\n## ")
         section = after[:next_heading] if next_heading != -1 else after
-        active_lines = [
-            line for line in section.splitlines()
-            if "<!-- autocontext:lesson-status=stale -->" not in line
-        ]
+        active_lines = [line for line in section.splitlines() if "<!-- autocontext:lesson-status=stale -->" not in line]
         return "\n".join(active_lines).strip()
 
     def write_hints(self, scenario_name: str, content: str, metadata: dict[str, Any] | None = None) -> None:
@@ -519,28 +526,12 @@ class ArtifactStore(PlaybookApprovalMethods, ArtifactGenerationPersistenceMethod
         records.sort(key=lambda record: (record.run_id, record.generation))
         return records
 
-    def harness_dir(self, scenario_name: str) -> Path:
-        """Return the harness directory: knowledge/<scenario>/harness/"""
-        return self._scenario_dir(scenario_name) / "harness"
-
-    @staticmethod
-    def _validate_harness_name(name: str) -> str:
-        """Validate harness module name and prevent path traversal."""
-        candidate = name.strip()
-        if not re.fullmatch(r"[a-zA-Z_][a-zA-Z0-9_]*", candidate):
-            raise ValueError(f"invalid harness name: {name!r}")
-        return candidate
-
     @staticmethod
     def _list_python_modules(directory: Path) -> list[str]:
         """List top-level Python modules, excluding private helper files."""
         if not directory.exists():
             return []
-        return sorted(
-            path.stem
-            for path in directory.glob("*.py")
-            if path.is_file() and not path.name.startswith("_")
-        )
+        return sorted(path.stem for path in directory.glob("*.py") if path.is_file() and not path.name.startswith("_"))
 
     @staticmethod
     def _render_python_context(
@@ -613,47 +604,6 @@ class ArtifactStore(PlaybookApprovalMethods, ArtifactGenerationPersistenceMethod
             created.append(f"{target.name} (updated)" if is_update else target.name)
 
         return created
-
-    def persist_harness(
-        self, scenario_name: str, generation_index: int, specs: list[dict[str, Any]],
-    ) -> list[str]:
-        """AST-validate and write harness .py files, archiving old versions."""
-        return self._persist_generated_modules(
-            self.harness_dir(scenario_name),
-            generation_index,
-            specs,
-            kind="harness",
-            header_template="Harness validator generated by architect in generation {generation_index}.",
-            name_validator=self._validate_harness_name,
-        )
-
-    def write_harness(self, scenario_name: str, name: str, source: str) -> Path:
-        """Write a single harness file to knowledge/<scenario>/harness/<name>.py."""
-        safe_name = self._validate_harness_name(name)
-        h_dir = self.harness_dir(scenario_name)
-        h_dir.mkdir(parents=True, exist_ok=True)
-        target = h_dir / f"{safe_name}.py"
-        target.write_text(source, encoding="utf-8")
-        return target
-
-    def read_harness(self, scenario_name: str, name: str) -> str | None:
-        """Read a harness file by name, or None if not found."""
-        safe_name = self._validate_harness_name(name)
-        target = self.harness_dir(scenario_name) / f"{safe_name}.py"
-        if not target.exists():
-            return None
-        return target.read_text(encoding="utf-8")
-
-    def list_harness(self, scenario_name: str) -> list[str]:
-        """List all harness file names for a scenario (sorted, without .py extension)."""
-        return self._list_python_modules(self.harness_dir(scenario_name))
-
-    def read_harness_context(self, scenario_name: str) -> str:
-        """Read harness validator files as markdown context for prompts."""
-        return self._render_python_context(
-            self.harness_dir(scenario_name),
-            empty_message="No harness validators available.",
-        )
 
     def tools_dir(self, scenario_name: str) -> Path:
         return self._scenario_dir(scenario_name) / "tools"
@@ -744,249 +694,6 @@ class ArtifactStore(PlaybookApprovalMethods, ArtifactGenerationPersistenceMethod
 
         return "\n\n".join(sections) if sections else "No generated tools available."
 
-    def persist_skill_note(self, scenario_name: str, generation_index: int, decision: str, lessons: str) -> None:
-        """Write a Claude Code Skill with playbook, lessons, and resource refs.
-
-        The skill directory becomes the knowledge hub for this scenario:
-
-        - ``SKILL.md`` — overview, lessons, and references (progressive disclosure)
-        - ``playbook.md`` — current consolidated strategy playbook (bundled resource)
-
-        Claude Code discovers the skill via YAML frontmatter and loads
-        ``SKILL.md`` on demand.  When deeper context is needed it reads
-        ``playbook.md`` (bundled) or follows references to the ``knowledge/``
-        directory for analysis history, tools, and raw coach output.
-        """
-        scenario = normalize_scenario_name_segment(scenario_name)
-        skill_dir = self._skill_dir(scenario)
-        skill_path = skill_dir / "SKILL.md"
-
-        existing_bullets: list[str] = []
-        if skill_path.exists():
-            in_lessons = False
-            for line in skill_path.read_text(encoding="utf-8").splitlines():
-                if line.startswith("## Operational Lessons"):
-                    in_lessons = True
-                    continue
-                if in_lessons and line.startswith("## "):
-                    break
-                if in_lessons and line.startswith("- "):
-                    existing_bullets.append(line)
-
-        if lessons and lessons.strip() not in ("", "No new lessons."):
-            for line in lessons.strip().splitlines():
-                stripped = line.strip()
-                if not stripped:
-                    continue
-                bullet = stripped if stripped.startswith("- ") else f"- {stripped}"
-                if bullet not in existing_bullets:
-                    existing_bullets.append(bullet)
-
-        kebab = scenario.replace("_", "-")
-        title = scenario.replace("_", " ").title()
-        desc = (
-            f"Operational knowledge for the {scenario} scenario including "
-            "strategy playbook, lessons learned, and resource references. "
-            f"Use when generating, evaluating, coaching, or debugging "
-            f"{scenario} strategies."
-        )
-        lessons_block = "\n".join(existing_bullets) if existing_bullets else "No lessons yet."
-
-        skill_content = (
-            f"---\nname: {kebab}-ops\ndescription: {desc}\n---\n\n"
-            f"# {title} Operational Knowledge\n\n"
-            "Accumulated knowledge from autocontext strategy evolution.\n\n"
-            "## Operational Lessons\n\n"
-            "Prescriptive rules derived from what worked and what failed:\n\n"
-            f"{lessons_block}\n\n"
-            "## Bundled Resources\n\n"
-            "- **Strategy playbook**: See [playbook.md](playbook.md) for the "
-            "current consolidated strategy guide (Strategy Updates, Prompt "
-            "Optimizations, Next Generation Checklist)\n"
-            f"- **Analysis history**: `knowledge/{scenario}/analysis/` "
-            "— per-generation analysis markdown\n"
-            f"- **Generated tools**: `knowledge/{scenario}/tools/` "
-            "— architect-created Python tools\n"
-            f"- **Coach history**: `knowledge/{scenario}/coach_history.md`"
-            " — raw coach output across all generations\n"
-            f"- **Architect changelog**: "
-            f"`knowledge/{scenario}/architect/changelog.md`"
-            " — infrastructure and tooling changes\n"
-        )
-
-        skill_dir.mkdir(parents=True, exist_ok=True)
-        skill_path.write_text(skill_content, encoding="utf-8")
-
-        playbook_content = self.read_playbook(scenario)
-        (skill_dir / "playbook.md").write_text(
-            playbook_content.strip() + "\n", encoding="utf-8",
-        )
-
-        self.sync_skills_to_claude()
-
-    def snapshot_knowledge(self, scenario_name: str, run_id: str) -> str:
-        """Copy playbook + skills + hints to snapshots/<run_id>/. Returns playbook hash."""
-        import hashlib
-
-        scenario_dir = self._scenario_dir(scenario_name)
-        snapshot_dir = scenario_dir / "snapshots" / run_id
-        snapshot_dir.mkdir(parents=True, exist_ok=True)
-
-        playbook_content = ""
-        playbook_path = scenario_dir / "playbook.md"
-        if playbook_path.exists():
-            playbook_content = playbook_path.read_text(encoding="utf-8")
-            (snapshot_dir / "playbook.md").write_text(playbook_content, encoding="utf-8")
-
-        hints_path = scenario_dir / "hints.md"
-        if hints_path.exists():
-            (snapshot_dir / "hints.md").write_text(
-                hints_path.read_text(encoding="utf-8"), encoding="utf-8"
-            )
-        hint_state_path = self._hint_state_path(scenario_name)
-        if hint_state_path.exists():
-            (snapshot_dir / "hint_state.json").write_text(
-                hint_state_path.read_text(encoding="utf-8"),
-                encoding="utf-8",
-            )
-
-        skill_dir = self._skill_dir(scenario_name)
-        skill_path = skill_dir / "SKILL.md"
-        if skill_path.exists():
-            (snapshot_dir / "SKILL.md").write_text(
-                skill_path.read_text(encoding="utf-8"), encoding="utf-8"
-            )
-
-        # Snapshot harness files
-        h_dir = self.harness_dir(scenario_name)
-        if h_dir.exists():
-            harness_snapshot = snapshot_dir / "harness"
-            harness_snapshot.mkdir(parents=True, exist_ok=True)
-            for py_file in h_dir.glob("*.py"):
-                if py_file.is_file():
-                    (harness_snapshot / py_file.name).write_text(
-                        py_file.read_text(encoding="utf-8"), encoding="utf-8",
-                    )
-
-        return hashlib.sha256(playbook_content.encode("utf-8")).hexdigest()[:16]
-
-    def restore_knowledge_snapshot(self, scenario_name: str, source_run_id: str) -> bool:
-        """Restore knowledge from a snapshot. Returns True if restored."""
-        scenario_dir = self._scenario_dir(scenario_name)
-        snapshot_dir = scenario_dir / "snapshots" / source_run_id
-        if not snapshot_dir.exists():
-            return False
-
-        restored = False
-        pb_snapshot = snapshot_dir / "playbook.md"
-        if pb_snapshot.exists():
-            self.write_playbook(scenario_name, pb_snapshot.read_text(encoding="utf-8"))
-            restored = True
-
-        hints_snapshot = snapshot_dir / "hints.md"
-        if hints_snapshot.exists():
-            self.write_markdown(
-                scenario_dir / "hints.md",
-                hints_snapshot.read_text(encoding="utf-8"),
-            )
-            restored = True
-        hint_state_snapshot = snapshot_dir / "hint_state.json"
-        if hint_state_snapshot.exists():
-            self.write_json(
-                self._hint_state_path(scenario_name),
-                read_json(hint_state_snapshot),
-            )
-            restored = True
-
-        skill_snapshot = snapshot_dir / "SKILL.md"
-        if skill_snapshot.exists():
-            skill_dir = self._skill_dir(scenario_name)
-            skill_dir.mkdir(parents=True, exist_ok=True)
-            (skill_dir / "SKILL.md").write_text(
-                skill_snapshot.read_text(encoding="utf-8"), encoding="utf-8"
-            )
-            restored = True
-
-        # Restore harness files from snapshot
-        harness_snapshot = snapshot_dir / "harness"
-        if harness_snapshot.exists():
-            h_dir = self.harness_dir(scenario_name)
-            h_dir.mkdir(parents=True, exist_ok=True)
-            for py_file in harness_snapshot.glob("*.py"):
-                if py_file.is_file():
-                    (h_dir / py_file.name).write_text(
-                        py_file.read_text(encoding="utf-8"), encoding="utf-8",
-                    )
-            restored = True
-
-        return restored
-
-    def read_skill_lessons_raw(self, scenario_name: str) -> list[str]:
-        """Return list of lesson bullet strings from SKILL.md."""
-        skill_path = self._skill_dir(scenario_name) / "SKILL.md"
-        if not skill_path.exists():
-            return []
-        bullets: list[str] = []
-        in_lessons = False
-        for line in skill_path.read_text(encoding="utf-8").splitlines():
-            if line.startswith("## Operational Lessons"):
-                in_lessons = True
-                continue
-            if in_lessons and line.startswith("## "):
-                break
-            if in_lessons and line.startswith("- "):
-                bullets.append(line)
-        return bullets
-
-    def replace_skill_lessons(self, scenario_name: str, lessons: list[str]) -> None:
-        """Replace the Operational Lessons section in SKILL.md with given bullets."""
-        skill_path = self._skill_dir(scenario_name) / "SKILL.md"
-        if not skill_path.exists():
-            return
-        content = skill_path.read_text(encoding="utf-8")
-        lines = content.splitlines()
-        result: list[str] = []
-        in_lessons = False
-        lessons_written = False
-        for line in lines:
-            if line.startswith("## Operational Lessons"):
-                result.append(line)
-                result.append("")
-                result.append("Prescriptive rules derived from what worked and what failed:")
-                result.append("")
-                for bullet in lessons:
-                    result.append(bullet if bullet.startswith("- ") else f"- {bullet}")
-                in_lessons = True
-                lessons_written = True
-                continue
-            if in_lessons:
-                if line.startswith("## "):
-                    in_lessons = False
-                    result.append("")
-                    result.append(line)
-                # Skip old lesson lines
-                continue
-            result.append(line)
-        if lessons_written:
-            skill_path.write_text("\n".join(result) + "\n", encoding="utf-8")
-
-    def sync_skills_to_claude(self) -> None:
-        """Symlink skill directories into .claude/skills/ for Claude Code discovery."""
-        self.claude_skills_path.mkdir(parents=True, exist_ok=True)
-        if not self.skills_root.exists():
-            return
-        for entry in self.skills_root.iterdir():
-            if not entry.is_dir() or not (entry / "SKILL.md").exists():
-                continue
-            link = self.claude_skills_path / entry.name
-            if link.is_symlink():
-                if link.resolve() == entry.resolve():
-                    continue
-                link.unlink()
-            elif link.exists():
-                continue  # Real file/dir exists, don't overwrite
-            os.symlink(entry.resolve(), link)
-
     def write_session_report(self, scenario_name: str, run_id: str, content: str) -> Path:
         """Write a session report for a completed run."""
         path = self._scenario_dir(scenario_name) / "reports" / f"{run_id}.md"
@@ -1028,12 +735,16 @@ class ArtifactStore(PlaybookApprovalMethods, ArtifactGenerationPersistenceMethod
     def _load_progress_report(self, data: dict[str, Any]) -> object:
         if data.get("schema_version") == 1 and "progress_points" in data:
             from autocontext.analytics.progress_report import RunProgressReport as CurveReport
+
             return CurveReport.from_dict(data)
         from autocontext.knowledge.normalized_metrics import RunProgressReport as NormalizedReport
+
         return NormalizedReport.from_dict(data)
 
     def read_latest_progress_reports(
-        self, scenario_name: str, max_reports: int = 2,
+        self,
+        scenario_name: str,
+        max_reports: int = 2,
     ) -> list[object]:
         """Read most recent progress reports for a scenario."""
         pr_dir = self._progress_report_dir(scenario_name)
@@ -1078,7 +789,9 @@ class ArtifactStore(PlaybookApprovalMethods, ArtifactGenerationPersistenceMethod
         return self._deserialize_weakness_report(data)
 
     def read_latest_weakness_reports(
-        self, scenario_name: str, max_reports: int = 2,
+        self,
+        scenario_name: str,
+        max_reports: int = 2,
     ) -> list[object]:
         """Read most recent weakness reports for a scenario."""
         wr_dir = self._weakness_dir(scenario_name)
@@ -1152,7 +865,11 @@ class ArtifactStore(PlaybookApprovalMethods, ArtifactGenerationPersistenceMethod
         return read_json(path)  # type: ignore[no-any-return]
 
     def _update_harness_version(
-        self, scenario_name: str, name: str, version: int, generation: int,
+        self,
+        scenario_name: str,
+        name: str,
+        version: int,
+        generation: int,
     ) -> None:
         versions = self.get_harness_version(scenario_name)
         versions[name] = {"version": version, "generation": generation}
@@ -1161,7 +878,11 @@ class ArtifactStore(PlaybookApprovalMethods, ArtifactGenerationPersistenceMethod
         write_json(path, versions)
 
     def write_harness_versioned(
-        self, scenario_name: str, name: str, source: str, generation: int,
+        self,
+        scenario_name: str,
+        name: str,
+        source: str,
+        generation: int,
     ) -> Path:
         """Write a harness file with version tracking, archiving the previous version."""
         normalized = self._validate_harness_name(name)
@@ -1188,7 +909,10 @@ class ArtifactStore(PlaybookApprovalMethods, ArtifactGenerationPersistenceMethod
         if isinstance(entry, dict) and isinstance(entry.get("version"), int) and entry["version"] > 1:
             entry["version"] -= 1
             self._update_harness_version(
-                scenario_name, normalized, entry["version"], entry.get("generation", 0),  # type: ignore[arg-type]
+                scenario_name,
+                normalized,
+                entry["version"],
+                entry.get("generation", 0),  # type: ignore[arg-type]
             )
         return store.read(filename)
